refactor: 패키지명 변경
- 대문자 오타 수정
