+++ src/main/java/bowling/domain/frame/FinalFrame.java
@@ -1,8 +1,8 @@
 package bowling.domain.frame;
 
-import bowling.domain.State.Pin;
-import bowling.domain.State.States;
 import bowling.domain.score.Score;
+import bowling.domain.state.Pin;
+import bowling.domain.state.States;
 
 import static bowling.view.OutputView.FINAL_STATE_FORMAT;
