+++ src/main/java/bowling/domain/state/First.java
@@ -1,26 +1,22 @@
-package bowling.domain.State;
+package bowling.domain.state;
 
 import bowling.domain.score.Score;
 
-public class Gutter extends RunningState {
-    private static final String SYMBOL = "-";
-
-    public Gutter() {
-        this(Pin.zero());
-    }
-
-    public Gutter(Pin pin) {
+public class First extends RunningState {
+    public First(Pin pin) {
         super(pin);
     }
 
     @Override
     public State bowl(Pin pin) {
-        if (pin.isTen()) {
-            return new Spare(pin, this);
-        }
+        Pin addedPin = this.pin.add(pin);
 
         if (pin.isZero()) {
             return new Miss(this);
+        }
+
+        if (addedPin.isTen()) {
+            return new Spare(addedPin, this);
         }
 
         return new Second(pin, this);
@@ -43,7 +39,7 @@ public class Gutter extends RunningState {
 
     @Override
     public String toString() {
-        return SYMBOL;
+        return pin.toString();
     }
 
     @Override
