+++ src/main/java/bowling/domain/state/Miss.java
@@ -1,4 +1,4 @@
-package bowling.domain.State;
+package bowling.domain.state;
 
 import bowling.domain.score.Score;
