+++ src/main/java/bowling/domain/state/RunningState.java
@@ -1,4 +1,4 @@
-package bowling.domain.State;
+package bowling.domain.state;
 
 public abstract class RunningState extends State {
     public RunningState(Pin pin) {
